kernel-shark: Add sched_waking event processing to sched_waking
It is now more common to enable the sched_waking event than to use the
sched_wakeup or sched_wakeup_new, as it encompasses both. But the sched
plugin does not handle that. Fix that by including the sched_waking event as
an event to use for processing schedule wakeup events.

Link: http://lore.kernel.org/linux-trace-devel/20190211012154.262506802@goodmis.org

Reviewed-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>