kernel-shark: Implement proper reser for the OpenGL widget
When loading a new trace.dat file the OpenGL widget has to be properly
reset. One important operation of this reset is to set the "_data"
pointer to NULL. This will block all attempts for drawing something
before the widget is set again. The bug was introduced in

7d057ae56 ("kernel-shark-qt: Add Trace Graph widget")

but had no effect until

94efea960 ("kernel-shark-qt: Handle the case when the range of ...")

because the number of bins of the model was zero, so nothing was plotted
anyway.

Link: http://lore.kernel.org/linux-trace-devel/20190214140451.827-1-ykaradzhov@vmware.com

Reported-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Tested-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Fixes: 7d057ae56 ("kernel-shark-qt: Add Trace Graph widget")
Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>