kernel-shark: Fix a bug in ksmodel_shift_backward()
Bin 0 of the new model has to be set after copying the overlaping bins
from the old model. Otherwise the new content of Bin 0 will be copied
into Bin "n". This bug has no effect because ksmodel_shift_backward()
has a second bug in the loop over the non-overlapping bins. The second
bug will be fixed in the following patch.

Link: http://lore.kernel.org/linux-trace-devel/20190221124205.21115-2-ykaradzhov@vmware.com

Reviewed-by: Slavomir Kaslev <kaslevs@vmware.com>
Fixes: f97e31f00 ("kernel-shark-qt: Introduce the visualization model ..")
Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>