kernel-shark: Disables automatic quotation for errors in KsMainWindow
When quoting in the output stream is disabled, the text of the
message is printed without quotations and without escaping of
non-printable characters.

example output with quoting enabled:
ERROR:  "Unable to open trace data file for session mysession.json\n"

and with quoting disabled:
ERROR:  Unable to open session description file mysession.json.

Link: http://lore.kernel.org/linux-trace-devel/20190314151012.905-6-ykaradzhov@vmware.com

Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>