kernel-shark: Avoid TRACECMD_BIN_DIR being set to the build path
The full path to the build directory (which is random for every user)
is embedded in the final binary. This is a no-no for some Linux distros
which are requiring that the finished binaries should not have such paths
embedded in them.

Link: http://lore.kernel.org/linux-trace-devel/20190614132609.16465-3-ykaradzhov@vmware.com
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=203187

Reported-by: Troy Engel <troyengel@gmail.com>
Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>