kernel-shark: kernel-shark: Avoid a duplicate error message from capture
If the capture process finishes with an error the _captureError() slot
will print an error message. We do not need to print identical message
in _captureFinished().

However, we have to keep in mind that the process that we started is
"pkexec" and the capture dialog is a child process of "pkexec". So if
the capture dialog failed to start or crashed the "pkexec" process
itself will have a normal exit, however its return value will be
different from 0 (it will just pass the return value of the capture
dialog executable). In this case _captureError() is not triggered,
but we still want to see an error message.

Link: http://lore.kernel.org/linux-trace-devel/20190710153906.32048-1-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
[ Fixed '||' to be '&&' as Yordan asked ]
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>