kernel-shark: Revert adding relative path to policy installation
This reverts the change in a45ab5a44 ("kernel-shark: Remove hard-coded
install paths") that removed the full path of installing
org.freedesktop.kshark-record.policy to a relative path. This means if we
installed in /usr/local/bin the policy still needs to go into
/usr/share/polkit-1/actions/

This does not affect DESTDIR, as installing still goes in the relative path
from DESTDIR for package maintainers.

Link: http://lore.kernel.org/linux-trace-devel/20190725194739.140392211@goodmis.org

Suggested-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Reviewed-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>