kernel-shark: Set text color to black in KsTraceGraph's axes labels
At present, background color of axes labels of the plot is forced to
white with a stylesheet, but the text color stays unchanged. As a
result, text is barely visible when using dark color schemes. This
patch forces the text color to black for these labels, making them
readable on all color schemes.

Link: https://lore.kernel.org/linux-trace-devel/20191010212429.469155-2-mike.rudenko@gmail.com/T/#u

Acked-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Mikhail Rudenko <mike.rudenko@gmail.com>
Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>