kernel-shark: Add stream interface for trace-cmd data
Here we provide an implementation of the Data stream interface that will
process the trace-cmd (Ftrace) data. This can be considered the most
essential change in the transformation of the C API towards version 2.
However, for the moment we only have stand alone definitions that are not
made functional yet. The actual integration with the API will be
introduced in the following patches.

Link: https://lore.kernel.org/linux-trace-devel/20201211150756.577366-14-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
[ Renamed tracefs_get_tracing_dir() to tracefs_tracing_dir() for libtracefs update ]
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>